Testing a couple of implementations of scanBlockString.
'Traditional' Bucketd approach is much more efficient in terms of memory consumption, but is slower because of how many conditions are required for it to work to spec. Using an approach more similar to the JavaScript implementation actually makes it much faster, but uses nearly 5x as much memory with with a fairly small block string value. I've also learnt that vektah's lexer doesn't handle a few things to spec, e.g. at the least, escape sequences (triple quotes). Trying to merge both solutions does produce a middleground in terms of memory usage and speed, (i.e. it gets a bit slower, but not as slow as the traditional approach, but uses less RAM, but not as little as the traditional approach).

More work needed here, there must be another approach to handling this without using byte buffers and string slices.